:art: Refactoring code
