+++ graphql/resolvers/index.js
@@ -1,0 +1,7 @@
+const usersResolvers = require("./users");
+
+module.exports = {
+  Mutation: {
+    ...usersResolvers.Mutation,
+  },
+};
